+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2022-08-22
+### Changed
+- Renamed `GenerateOutputBroadcasters` to `GenerateOutputBroadcast`
+- Renamed `GenerateInputOutputBroadcasters` to `GenerateBroadcast`
+
 ## [0.2.0] - 2022-08-21
 ### Added
 - Examples in go documentation for [pkg.go.dev/github.com/JamesChung/fofi](https://pkg.go.dev/github.com/JamesChung/fofi)